refactor: Panel 컴포넌트 분리
