+++ src/utils/shop/handlePrice.ts
@@ -1,5 +1,5 @@
 
-const handlePrice = (price: number) => {
+const handlePrice = (price: number | string) => {
   return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
 }
